+++ eval/tests/check_readme.rs
@@ -16,6 +16,7 @@ fn check_sample(code_sample: &str) {
     let program = Untyped::<F32Grammar>::parse_statements(code_sample).unwrap();
 
     let mut env: Environment<'_, f32> = Prelude.iter().chain(Assertions.iter()).collect();
+    env.insert("INF", Value::Prim(f32::INFINITY));
     env.insert("array", Value::native_fn(fns::Array)).insert(
         "rand_num",
         Value::wrapped_fn(|min: f32, max: f32| thread_rng().gen_range(min..max)),